+++ src/Blocks/components/submit/components/submit-editor.js
@@ -4,9 +4,9 @@ import { FieldEditor } from '../../../components/field/components/field-editor';
 import manifest from '../manifest.json';
 
 export const SubmitEditor = (attributes) => {
-	const submitValue = checkAttr('submitValue', attributes, manifest) || __('Submit', 'eightshift-forms');
+	const submitValue = checkAttr('submitValue', attributes, manifest);
 
-	const submitComponent = <button className='esf:w-full esf:p-10 esf:rounded-md esf:text-sm esf:text-white esf:bg-mist-600 esf:font-bold'>{submitValue}</button>;
+	const submitComponent = <button className='esf:w-full esf:p-10 esf:rounded-md esf:text-sm esf:text-white esf:bg-mist-600 esf:font-bold'>{submitValue ? submitValue : __('Submit', 'eightshift-forms')}</button>;
 
 	return (
 		<>