+++ src/CronJobs/EntriesAutoDeleteJob.php
@@ -43,13 +43,8 @@ class EntriesAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 	 */
 	public function register(): void
 	{
-		if (!\apply_filters(SettingsEntries::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
-			return;
-		}
-
-		\add_action('admin_init', $this->checkIfJobIsSet(...));
-		\add_filter('cron_schedules', $this->addJobToSchedule(...)); // phpcs:ignore WordPress.WP.CronInterval.ChangeDetected
-		\add_action(self::JOB_NAME, $this->getJobCallback(...));
+		\add_action('admin_init', [$this, 'checkIfJobIsSet']);
+		\add_action(self::JOB_NAME, [$this, 'getJobCallback']);
 	}
 
 	/**
@@ -60,27 +55,10 @@ class EntriesAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 		if (!\wp_next_scheduled(self::JOB_NAME)) {
 			\wp_schedule_event(
 				\strtotime('tomorrow', \time()),
-				'daily',
+				CronJobsSchedules::CRON_JOBS_SCHEDULE_EVERY_DAY,
 				self::JOB_NAME
 			);
 		}
-	}
-
-	/**
-	 * Add job to schedule.
-	 *
-	 * @param array<mixed> $schedules WP schedules list.
-	 *
-	 * @return array<mixed>
-	 */
-	public function addJobToSchedule(array $schedules): array
-	{
-		$schedules['daily'] = [
-			'interval' => \DAY_IN_SECONDS,
-			'display' => \esc_html__('Every day at midnight', 'eightshift-forms'),
-		];
-
-		return $schedules;
 	}
 
 	/**
@@ -88,6 +66,10 @@ class EntriesAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 	 */
 	public function getJobCallback(): void
 	{
+		if (!\apply_filters(SettingsEntries::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
+			return;
+		}
+
 		$forms = $this->formsListing->getFormsList([
 			'post_type' => Forms::POST_TYPE_SLUG,
 		]);