+++ src/CronJobs/LogEntryCleanupJob.php
@@ -31,13 +31,8 @@ class LogEntryCleanupJob implements ServiceInterface
 	 */
 	public function register(): void
 	{
-		if (!\apply_filters(SettingsSecurity::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
-			return;
-		}
-
-		\add_action('init', $this->checkIfJobIsSet(...));
-		\add_filter('cron_schedules', $this->addJobToSchedule(...)); // phpcs:ignore WordPress.WP.CronInterval.ChangeDetected
-		\add_action(self::JOB_NAME, $this->getJobCallback(...));
+		\add_action('init', [$this, 'checkIfJobIsSet']);
+		\add_action(self::JOB_NAME, [$this, 'getJobCallback']);
 	}
 
 	/**
@@ -48,27 +43,10 @@ class LogEntryCleanupJob implements ServiceInterface
 		if (!\wp_next_scheduled(self::JOB_NAME)) {
 			\wp_schedule_event(
 				\strtotime('tomorrow', \time()),
-				'daily',
+				CronJobsSchedules::CRON_JOBS_SCHEDULE_EVERY_DAY,
 				self::JOB_NAME
 			);
 		}
-	}
-
-	/**
-	 * Add job to schedule.
-	 *
-	 * @param array<mixed> $schedules WP schedules list.
-	 *
-	 * @return array<mixed>
-	 */
-	public function addJobToSchedule(array $schedules): array
-	{
-		$schedules['daily'] = [
-			'interval' => \DAY_IN_SECONDS,
-			'display' => \esc_html__('Every day at midnight', 'eightshift-forms'),
-		];
-
-		return $schedules;
 	}
 
 	/**
@@ -76,6 +54,10 @@ class LogEntryCleanupJob implements ServiceInterface
 	 */
 	public function getJobCallback(): void
 	{
+		if (!\apply_filters(SettingsSecurity::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
+			return;
+		}
+
 		RateLimitingLogEntry::cleanup(\DAY_IN_SECONDS);
 	}
 }