+++ src/CronJobs/NationbuilderJob.php
@@ -57,13 +57,8 @@ class NationbuilderJob implements ServiceInterface, ServiceCliInterface
 	 */
 	public function register(): void
 	{
-		if (!\apply_filters(SettingsNationbuilder::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
-			return;
-		}
-
-		\add_action('admin_init', $this->checkIfJobIsSet(...));
-		\add_filter('cron_schedules', $this->addJobToSchedule(...)); // phpcs:ignore WordPress.WP.CronInterval.CronSchedulesInterval
-		\add_action(self::JOB_NAME, $this->getJobCallback(...));
+		\add_action('admin_init', [$this, 'checkIfJobIsSet']);
+		\add_action(self::JOB_NAME, [$this, 'getJobCallback']);
 	}
 
 	/**
@@ -73,28 +68,11 @@ class NationbuilderJob implements ServiceInterface, ServiceCliInterface
 	{
 		if (!\wp_next_scheduled(self::JOB_NAME)) {
 			\wp_schedule_event(
-				\time(),
-				'esFormsEvery5Minutes',
+				\strtotime('tomorrow', \time()),
+				CronJobsSchedules::CRON_JOBS_SCHEDULE_EVERY_15_MINUTES,
 				self::JOB_NAME
 			);
 		}
-	}
-
-	/**
-	 * Add job to schedule.
-	 *
-	 * @param array<mixed> $schedules WP schedules list.
-	 *
-	 * @return array<mixed>
-	 */
-	public function addJobToSchedule(array $schedules): array
-	{
-		$schedules['esFormsEvery5Minutes'] = [
-			'interval' => \MINUTE_IN_SECONDS * 5,
-			'display' => \esc_html__('Every 5 minutes', 'eightshift-forms'),
-		];
-
-		return $schedules;
 	}
 
 	/**
@@ -162,8 +140,5 @@ class NationbuilderJob implements ServiceInterface, ServiceCliInterface
 		}
 
 		\update_option(SettingsHelpers::getOptionName(SettingsNationbuilder::SETTINGS_NATIONBUILDER_CRON_KEY), $jobs);
-
-		// Turn of OAuth after cron job is done.
-		\delete_option(SettingsHelpers::getOptionName(SettingsNationbuilder::SETTINGS_NATIONBUILDER_OAUTH_ALLOW_KEY));
 	}
 }