+++ src/Rest/Routes/Integrations/Pardot/FormSubmitPardotRoute.php
@@ -1,0 +1,153 @@
+<?php
+
+/**
+ * Form submit route for Pardot.
+ *
+ * @package EightshiftForms\Rest\Routes\Integrations\Pardot
+ */
+
+declare(strict_types=1);
+
+namespace EightshiftForms\Rest\Routes\Integrations\Pardot;
+
+use EightshiftForms\Captcha\CaptchaInterface;
+use EightshiftForms\Enrichment\EnrichmentInterface;
+use EightshiftForms\Integrations\Pardot\PardotClientInterface;
+use EightshiftForms\Integrations\Pardot\SettingsPardot;
+use EightshiftForms\Labels\LabelsInterface;
+use EightshiftForms\Integrations\Mailer\MailerInterface;
+use EightshiftForms\Rest\Routes\AbstractIntegrationFormSubmit;
+use EightshiftForms\Security\SecurityInterface;
+use EightshiftForms\Validation\ValidatorInterface;
+use EightshiftForms\Config\Config;
+use EightshiftForms\Exception\BadRequestException;
+use EightshiftForms\Exception\DisabledIntegrationException;
+use EightshiftForms\Helpers\SettingsHelpers;
+use EightshiftForms\Rest\Routes\AbstractBaseRoute;
+use EightshiftForms\Troubleshooting\SettingsFallback;
+
+/**
+ * Class FormSubmitPardotRoute
+ */
+class FormSubmitPardotRoute extends AbstractIntegrationFormSubmit
+{
+	/**
+	 * Route slug.
+	 */
+	public const ROUTE_SLUG = SettingsPardot::SETTINGS_TYPE_KEY;
+
+	/**
+	 * Instance variable for Pardot data.
+	 *
+	 * @var PardotClientInterface
+	 */
+	protected $pardotClient;
+
+	/**
+	 * Create a new instance that injects classes
+	 *
+	 * @param SecurityInterface $security Inject security methods.
+	 * @param ValidatorInterface $validator Inject validator methods.
+	 * @param LabelsInterface $labels Inject labels methods.
+	 * @param CaptchaInterface $captcha Inject captcha methods.
+	 * @param MailerInterface $mailer Inject mailerInterface methods.
+	 * @param EnrichmentInterface $enrichment Inject enrichment methods.
+	 * @param PardotClientInterface $pardotClient Inject pardotClient methods.
+	 */
+	public function __construct(
+		SecurityInterface $security,
+		ValidatorInterface $validator,
+		LabelsInterface $labels,
+		CaptchaInterface $captcha,
+		MailerInterface $mailer,
+		EnrichmentInterface $enrichment,
+		PardotClientInterface $pardotClient
+	) {
+		$this->security = $security;
+		$this->validator = $validator;
+		$this->labels = $labels;
+		$this->captcha = $captcha;
+		$this->mailer = $mailer;
+		$this->enrichment = $enrichment;
+		$this->pardotClient = $pardotClient;
+	}
+
+	/**
+	 * Get the base url of the route
+	 *
+	 * @return string The base URL for route you are adding.
+	 */
+	protected function getRouteName(): string
+	{
+		return '/' . Config::ROUTE_PREFIX_FORM_SUBMIT . '/' . self::ROUTE_SLUG;
+	}
+
+	/**
+	 * Check if the route is admin protected.
+	 *
+	 * @return boolean
+	 */
+	protected function isRouteAdminProtected(): bool
+	{
+		return false;
+	}
+
+	/**
+	 * Get mandatory params.
+	 *
+	 * @param array<string, mixed> $params Params passed from the request.
+	 *
+	 * @return array<string, string>
+	 */
+	protected function getMandatoryParams(array $params): array
+	{
+		return [
+			Config::FD_FORM_ID => 'string',
+			Config::FD_POST_ID => 'string',
+			Config::FD_PARAMS => 'array',
+		];
+	}
+
+	/**
+	 * Implement submit action.
+	 *
+	 * @param array<string, mixed> $formDetails Data passed from the `getFormDetailsApi` function.
+	 *
+	 * @throws BadRequestException If Pardot is missing config.
+	 * @throws DisabledIntegrationException If Pardot is disabled.
+	 *
+	 * @return mixed
+	 */
+	protected function submitAction(array $formDetails)
+	{
+		if (SettingsHelpers::isOptionCheckboxChecked(SettingsPardot::SETTINGS_PARDOT_SKIP_INTEGRATION_KEY, SettingsPardot::SETTINGS_PARDOT_SKIP_INTEGRATION_KEY)) {
+			$integrationSuccessResponse = $this->getIntegrationResponseSuccessOutput($formDetails);
+
+			// phpcs:disable Eightshift.Security.HelpersEscape.ExceptionNotEscaped
+			throw new DisabledIntegrationException(
+				$integrationSuccessResponse[AbstractBaseRoute::R_MSG],
+				$integrationSuccessResponse[AbstractBaseRoute::R_DEBUG],
+				$integrationSuccessResponse[AbstractBaseRoute::R_DATA]
+			);
+			// phpcs:enable
+		}
+
+		if (!\apply_filters(SettingsPardot::FILTER_SETTINGS_IS_VALID_NAME, false, $formDetails[Config::FD_FORM_ID])) {
+			// phpcs:disable Eightshift.Security.HelpersEscape.ExceptionNotEscaped
+			throw new BadRequestException(
+				$this->getLabels()->getLabel('pardotMissingConfig'),
+				[
+					AbstractBaseRoute::R_DEBUG => $formDetails,
+					AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_PARDOT_MISSING_CONFIG,
+				],
+			);
+			// phpcs:enable
+		}
+
+		$response = $this->pardotClient->postApplication($formDetails);
+
+		$formDetails[Config::FD_RESPONSE_OUTPUT_DATA] = $response;
+
+		return $this->getIntegrationCommonSubmitAction($formDetails);
+	}
+}